перевод в double

+++ wpfCalc/MainWindow.xaml.cs
@@ -20,8 +20,8 @@ namespace wpfCalc
     /// </summary>
     public partial class MainWindow : Window
     {
-        int num1 = 0;
-        int num2 = 0;
+        double num1 = 0;
+        double num2 = 0;
         string op = "";
         public MainWindow()
         {
@@ -53,7 +53,7 @@ namespace wpfCalc
 
         private void btn_eq_Click(object sender, RoutedEventArgs e)
         {
-            int result = 0;
+            double result = 0;
 
             switch (op)
             {
@@ -79,7 +79,7 @@ namespace wpfCalc
                     result = num1 > num2 ? num2 : num1;
                     break;
                 case "x^y":
-                    result = Pow(num1, num2);
+                    result = Pow(num1, (int)num2);
                     break;
             }
             txtBox.Text = result.ToString();
@@ -89,7 +89,7 @@ namespace wpfCalc
 
         }
 
-        private int Pow(int x, int y)
+        private double Pow(double x, int y)
         {
             if (y == 0)
                 return 1;
